EditDialog: rename ClimbingMultiValues to ClimbingEditor
