+++ src/components/FeaturePanel/EditDialog/EditContent/FeatureEditSection/ClimbingEditor/EditorHeader.tsx
