+++ src/components/FeaturePanel/EditDialog/EditContent/FeatureEditSection/ClimbingEditor/EditorItem.tsx
